pyproject.toml
[project]
name = "ghostops"
version = "1.0.0"
description = "Gemini Live API desktop agent with persistent cross-session memory"
requires-python = ">=3.12"
dependencies = [
    # AI + Cloud
    "google-genai>=0.8.0",
    "google-cloud-firestore>=2.19.0",
    # Backend (Cloud Run)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "google-adk>=1.27.1",
    "google-api-core>=2.29.0",
    "pyaudio>=0.2.14",
    "mss>=10.1.0",
    "pillow>=12.1.0",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
    "keyboard>=0.13.5",
    "numpy>=2.4.3",
    "browser-use>=0.12.2",
    "playwright>=1.58.0",
    "requests>=2.32.5",
    "python-vlc>=3.0.21203",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
desktop = [
    # Only needed on the user's machine, not on Cloud Run
    "pyaudio>=0.2.14",
    "mss>=9.0.2",
    "Pillow>=10.0.0",
    "pyautogui>=0.9.54",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
ghostops = "desktop.main:main"
ghostops-ptt = "desktop.ptt:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
markers = [
    "integration: tests that hit live APIs (Gemini, Firestore, Cloud Run)",
]
